chore: Rename interactives to interactive
