+++ packages/interactive/.eslintignore
