+++ packages/interactive/.npmignore
