+++ packages/interactive/__tests__/web/jest.config.js
