+++ packages/interactive/interactive.stories.js
@@ -1,4 +1,4 @@
 import { showcaseConverter } from "@times-components/storybook";
-import showcase from "./interactives.showcase";
+import showcase from "./interactive.showcase";
 
 showcaseConverter(module, showcase);
