+++ packages/interactive/webpack.config.js
